% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{analyzeCE}
\alias{analyzeCE}
\title{Compare Cost-Effectiveness strategies}
\usage{
analyzeCE(
  ...,
  context = list(),
  cost.label = NULL,
  eff.label = NULL,
  nudge.x.icer = -0.03,
  nudge.y.icer = 0.02
)
}
\arguments{
\item{...}{CEAModel list of the strategies to be evaluated and compared}

\item{context}{context with values of variables}
}
\value{
list with keys: 'summary' (comparison table) and 'plot' (efficiency curve)
}
\description{
Compares cost-effectiveness of different models, generating a comparison table and
a efficiency curve plot.
}
\examples{
library(CEAModel)
tree1 <- loadDecisionTree('tree1.yaml')
tree2 <- loadDecisionTree('tree2.yaml')
tree3 <- loadDecisionTree('tree3.yaml')
ctx <- list( ... )
outputs <- analyzeCE(tree1=tree1, tree2=tree2, tree3=tree3, context=ctx)

}
