% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionTree.R
\docType{class}
\name{DecisionTree-class}
\alias{DecisionTree-class}
\alias{DecisionTree}
\title{Decision Tree for Cost-Effectiveness Analysis}
\description{
Decision Tree for Cost-Effectiveness Analysis
}

