% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{generateContextTemplate}
\alias{generateContextTemplate}
\title{Generates a stratified context XLSX file}
\usage{
generateContextTemplate(..., filename, strata = "base")
}
\arguments{
\item{...}{CEAModel(s)}

\item{filename}{Filename of the xlsx file}

\item{strata}{Names of the strata}
}
\description{
Generates a stratified context XLSX file for the parameters required by the specified CEA models.
}
\examples{
library(CEAModel)
tree <- loadDecisionTree('tree.yaml')
markov <- loadMarkovModel('markov.xlsx')
generateContextTemplate(tree, markov, 'context.xlsx')

}
