% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{generateEmptyContextFromModels}
\alias{generateEmptyContextFromModels}
\title{Generates a context list from CEA models}
\usage{
generateEmptyContextFromModels(...)
}
\arguments{
\item{...}{CEAModel(s)}
}
\description{
Generates a context list from CEA models such as decision trees, markov or microsimulation models,
all of them initialized to 0.
}
\examples{
library(CEAModel)
tree <- loadDecisionTree('tree.yaml')
markov <- loadMarkovModel('markov.xlsx')
generateEmptyContextFromModels(tree, markov)

}
