% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{getIndependentParameters}
\alias{getIndependentParameters}
\title{Get all independent parameter names from context XLSX file}
\usage{
getIndependentParameters(filename)
}
\arguments{
\item{filename}{Filename of the xlsx file}
}
\description{
Returns the names of all parameters not dependent on other values (i.e. excel cells).
}
\examples{
library(CEAModel)
ctx.filename <- 'context.xlsx'
ctx <- loadStratifiedContextFile(ctx.filename)
independent.params <- getIndependentParameters(ctx.filename)
}
