% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{loadContextFile}
\alias{loadContextFile}
\title{Load a XLSX file to a context list for CEA models}
\usage{
loadContextFile(filename, sheet = 1)
}
\arguments{
\item{filename}{Filename of the xlsx file}
}
\value{
list with context variables
}
\description{
Generates a context list for CEA models from an excel file.
}
\examples{
library(CEAModel)
ctx <- loadContextFile('context.xlsx')
tree1 <- loadDecisionTree('tree1.yaml')
tree2 <- loadDecisionTree('tree2.yaml')
outputs <- analyzeCE(tree1=tree1, tree2=tree2, context=ctx)

}
