% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionTree.R
\name{loadDecisionTree}
\alias{loadDecisionTree}
\title{Load decision tree from YAML file}
\usage{
loadDecisionTree(filePath)
}
\arguments{
\item{filePath}{path to YAML file with the tree specification}
}
\value{
DecisionTree
}
\description{
Generates a decision tree as specified by the YAML file.
}
\details{
TODO: Document tree specification format.
}
\examples{
library(CEAModel)
tree <- loadDecisionTree('example.yaml')
print(tree)

}
