% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{loadMarkovModels}
\alias{loadMarkovModels}
\title{Load markov model from XLSX file}
\usage{
loadMarkovModels(filePaths)
}
\arguments{
\item{filePath}{path to XLSX file with the markov model specification}
}
\value{
MarkovModel
}
\description{
Generates a markov model as specified by the XLSX file.
}
\details{
TODO: Document markov model specification format.
}
\examples{
library(CEAModel)
markov <- loadMarkovModels('example.xlsx')
print(markov)

}
