% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimulation.R
\name{loadMicrosimulationModel}
\alias{loadMicrosimulationModel}
\title{Load microsimulation model from XLSX file and simulation functions}
\usage{
loadMicrosimulationModel(
  markov.file.paths,
  setup.individuals.func,
  aggregate.individual.results.func,
  aggregate.population.results.func,
  sanity.check.func = NULL
)
}
\arguments{
\item{markovFilePaths}{path to XLSX file(s) with the markov model specification(s)}

\item{setupIndividualsFunc}{function that sets up the initial state for the simulated subjects}

\item{aggregateIndividualResultsFunc}{function that aggregates the results from an entire subject simulation history into an individual summary}

\item{aggregatePopulationResultsFunc}{function that aggregates the results from all individual summaries into a global summary.}

\item{sanityCheckFunc}{function that checks whether the context variables are valid or not. By default it returns always true (no check performed)}
}
\value{
MicrosimulationModel
}
\description{
Generates a microsimulation model as specified by the XLSX file (markov models) and the following necessary functions.
}
\examples{
TODO: Finish example
library(CEAModel)
simulateStep <- function(subject, t, context) {}
aggregateIndividualResults <- function(results, context) {}
aggregatePopulationResults <- function(results, context) {}
calibrate <- function(expectedOutput, context) {}
sanityCheck = function(context) {}
ms.model <- loadMicrosimulationModel(markovFilePaths='example.xlsx',
                                     setupIndividualsFunc = setupIndividuals,
                                     aggregateIndividualResultsFunc = aggregateIndividualResults,
                                     aggregatePopulationResultsFunc = aggregatePopulationResults,
                                     sanityCheckFunc = sanityCheck
                                     )
print(ms.model)

}
