% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{loadStratifiedContextFile}
\alias{loadStratifiedContextFile}
\title{Load a XLSX file to a stratified context list for CEA models}
\usage{
loadStratifiedContextFile(filename)
}
\arguments{
\item{filename}{Filename of the xlsx file}
}
\value{
list of strata, each with context variables
}
\description{
Generates a stratified context list for CEA models from an excel file.
}
\examples{
library(CEAModel)
ctx <- loadStratifiedContextFile('context.xlsx')
tree1 <- loadDecisionTree('tree1.yaml')
tree2 <- loadDecisionTree('tree2.yaml')
outputs1 <- analyzeCE(tree1=tree1, tree2=tree2, context=ctx$stratum1)
outputs2 <- analyzeCE(tree1=tree1, tree2=tree2, context=ctx$stratum2)

}
