% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsimulation.R
\docType{class}
\name{MicrosimulationModel-class}
\alias{MicrosimulationModel-class}
\alias{MicrosimulationModel}
\title{Microsimulation Model for Cost-Effectiveness Analysis}
\description{
It specifies and specifies and run a microsimulation model. For more information about the functions required, see Details.
}
\details{

}

