% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{saveContextFile}
\alias{saveContextFile}
\title{Save a context list for CEA models to a XLSX file}
\usage{
saveContextFile(context, filename)
}
\arguments{
\item{context}{Named list with context variables and their values}

\item{filename}{Filename of the xlsx file}
}
\description{
Generates an excel file from a context list for CEA models.
}
\examples{
library(CEAModel)
ctx <- list(var1=12, var2=34)
saveContextFile(ctx, 'context.xlsx')

}
