% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{saveMarkovModel}
\alias{saveMarkovModel}
\title{Save markov model to XLSX file}
\usage{
saveMarkovModel(markov, filePath)
}
\arguments{
\item{markov}{markov model (or named list of markov models)}

\item{filePath}{path to XLSX file with the markov model specification}
}
\description{
Generates a XLSX file from a markov model (or list of markov models).
}
\details{
TODO: Document markov model specification format.
}
\examples{
library(CEAModel)
markov <- loadMarkovModels('example.xlsx')
...
saveMarkovModel(markov, 'example.modified.xlsx')

}
