% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{saveStratifiedContextFile}
\alias{saveStratifiedContextFile}
\title{Save a stratified context list for CEA models to a XLSX file}
\usage{
saveStratifiedContextFile(strat.context, filename)
}
\arguments{
\item{strat.context}{Named list with strata, each of them with context variables and their values}

\item{filename}{Filename of the xlsx file}
}
\description{
Generates an excel file from a stratified context list for CEA models.
}
\examples{
library(CEAModel)
ctx <- list(
        stratum1=list(var1=12, var2=34),
        stratum2=list(var1=23, var2=45)
       )
saveStratifiedContextFile(ctx, 'context.xlsx')

}
